refactor: 카테고리 패키지 구조 변경
