+++ src/main/java/com/sparta/todayeats/category/dto/request/CategoryUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.sparta.todayeats.category.presentation.dto;
+package com.sparta.todayeats.category.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
