+++ src/main/java/com/sparta/todayeats/category/dto/response/CategoryResponse.java
@@ -1,4 +1,4 @@
-package com.sparta.todayeats.category.presentation.dto;
+package com.sparta.todayeats.category.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -6,14 +6,17 @@ import lombok.Getter;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
-// 카테고리 생성 응답 DTO
+// 카테고리 정보 응답 DTO
 @Getter
 @Builder
-public class CategoryCreateResponse {
+public class CategoryResponse {
 
     private UUID categoryId;
     private String name;
 
     private LocalDateTime createdAt;
     private UUID createdBy;
+
+    private LocalDateTime updatedAt;
+    private UUID updatedBy;
 }
