+++ src/main/java/com/sparta/todayeats/store/entity/Store.java
@@ -1,7 +1,7 @@
 package com.sparta.todayeats.store.entity;
 
 import com.sparta.todayeats.area.domain.entity.Area;
-import com.sparta.todayeats.category.domain.entity.Category;
+import com.sparta.todayeats.category.entity.Category;
 import com.sparta.todayeats.global.infrastructure.entity.BaseEntity;
 import com.sparta.todayeats.user.entity.User;
 import jakarta.persistence.*;
