+++ src/main/java/com/sparta/todayeats/store/service/StoreService.java
@@ -2,8 +2,8 @@ package com.sparta.todayeats.store.service;
 
 import com.sparta.todayeats.area.domain.entity.Area;
 import com.sparta.todayeats.area.domain.repository.AreaRepository;
-import com.sparta.todayeats.category.domain.entity.Category;
-import com.sparta.todayeats.category.domain.repository.CategoryRepository;
+import com.sparta.todayeats.category.entity.Category;
+import com.sparta.todayeats.category.repository.CategoryRepository;
 import com.sparta.todayeats.global.response.PageResponse;
 import com.sparta.todayeats.global.exception.*;
 import com.sparta.todayeats.store.dto.request.StoreCreateRequest;
